feat(product/position) : 포지션 기능 구현
