+++ src/main/java/com/nhnacademy/front/product/contributor/position/dto/response/ResponsePositionDTO.java
@@ -1,18 +1,19 @@
-package com.nhnacademy.front.product.contributor.model.dto.request;
+package com.nhnacademy.front.product.contributor.position.dto.response;
 
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class RequestPositionDTO {
+public class ResponsePositionDTO {
+	@NotNull
+	private long positionId;
 	@NotBlank
 	private String positionName;
-
-	public void setPositionName(String positionName) {
-		this.positionName = positionName;
-	}
 }
