+++ src/main/java/com/nhnacademy/front/product/contributor/position/exception/PositionGetProcessException.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.front.product.contributor.exception;
+package com.nhnacademy.front.product.contributor.position.exception;
 
 public class PositionGetProcessException extends RuntimeException {
 	public PositionGetProcessException(String message) {
