+++ src/main/java/com/nhnacademy/front/product/contributor/position/exception/PositionProcessException.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.front.product.contributor.exception;
+package com.nhnacademy.front.product.contributor.position.exception;
 
 public class PositionProcessException extends RuntimeException {
 	public PositionProcessException(String message) {
