+++ src/main/java/com/nhnacademy/front/product/contributor/position/service/PositionService.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.front.product.contributor.service;
+package com.nhnacademy.front.product.contributor.position.service;
 
 import java.util.Objects;
 
@@ -8,11 +8,11 @@ import org.springframework.stereotype.Service;
 
 import com.nhnacademy.front.common.exception.EmptyRequestException;
 import com.nhnacademy.front.common.page.PageResponse;
-import com.nhnacademy.front.product.contributor.adaptor.PositionAdaptor;
-import com.nhnacademy.front.product.contributor.exception.PositionGetProcessException;
-import com.nhnacademy.front.product.contributor.exception.PositionProcessException;
-import com.nhnacademy.front.product.contributor.model.dto.request.RequestPositionDTO;
-import com.nhnacademy.front.product.contributor.model.dto.response.ResponsePositionDTO;
+import com.nhnacademy.front.product.contributor.position.adaptor.PositionAdaptor;
+import com.nhnacademy.front.product.contributor.position.exception.PositionGetProcessException;
+import com.nhnacademy.front.product.contributor.position.exception.PositionProcessException;
+import com.nhnacademy.front.product.contributor.position.dto.request.RequestPositionDTO;
+import com.nhnacademy.front.product.contributor.position.dto.response.ResponsePositionDTO;
 
 import feign.FeignException;
 import lombok.RequiredArgsConstructor;
