+++ src/main/resources/application.yml
@@ -2,6 +2,10 @@
 spring:
   application:
     name: FRONT
+  mvc:
+    hiddenmethod:
+      filter:
+        enabled: true
 
 logging:
   level:
@@ -16,3 +20,4 @@ management:
     web:
       exposure:
         include: "*" # Actuator 의 모든 엔드포인트를 웹에서 노출 (운영환경에선 최소화 필요)
+
